+++ src/ts/create-cards-value.ts
@@ -26,6 +26,19 @@ export const cardSuit = [
     },
 ];
 
+function createObjCard(
+    value: string,
+    cardSuits: CardSuit[],
+    index: number,
+    arrayCard: Card[]
+) {
+    const infoAboutCardSpades: Card = {};
+    infoAboutCardSpades.cardValue = value;
+    infoAboutCardSpades.cardImage = cardSuits[index].image;
+    infoAboutCardSpades.cardName = `${value}-${cardSuits[0].color}`;
+    arrayCard.push(infoAboutCardSpades);
+}
+
 export function createCardValues(
     cardSuits: CardSuit[],
     cardValues: string[]
@@ -33,29 +46,10 @@ export function createCardValues(
     const arrayCard: Card[] = [];
 
     cardValues.forEach((value) => {
-        const infoAboutCardSpades: Card = {};
-        infoAboutCardSpades.cardValue = value;
-        infoAboutCardSpades.cardImage = cardSuits[0].image;
-        infoAboutCardSpades.cardName = `${value}-${cardSuits[0].color}`;
-        arrayCard.push(infoAboutCardSpades);
-
-        const infoAboutCardHearts: Card = {};
-        infoAboutCardHearts.cardValue = value;
-        infoAboutCardHearts.cardImage = cardSuits[1].image;
-        infoAboutCardHearts.cardName = `${value}-${cardSuits[1].color}`;
-        arrayCard.push(infoAboutCardHearts);
-
-        const infoAboutCardDiamonds: Card = {};
-        infoAboutCardDiamonds.cardValue = value;
-        infoAboutCardDiamonds.cardImage = cardSuits[2].image;
-        infoAboutCardDiamonds.cardName = `${value}-${cardSuits[2].color}`;
-        arrayCard.push(infoAboutCardDiamonds);
-
-        const infoAboutCardClubs: Card = {};
-        infoAboutCardClubs.cardValue = value;
-        infoAboutCardClubs.cardImage = cardSuits[3].image;
-        infoAboutCardClubs.cardName = `${value}-${cardSuits[3].color}`;
-        arrayCard.push(infoAboutCardClubs);
+        createObjCard(value, cardSuits, 0, arrayCard);
+        createObjCard(value, cardSuits, 1, arrayCard);
+        createObjCard(value, cardSuits, 2, arrayCard);
+        createObjCard(value, cardSuits, 3, arrayCard);
     });
 
     return arrayCard;
