+++ src/ts/template-engine/create-template-card.ts
@@ -1,7 +1,7 @@
 import { HtmlEl } from '../types';
 
 export function createTempleteCart(
-    block: HtmlEl | HtmlEl[] | string | (HtmlEl | string)[] | undefined
+    block?: HtmlEl | HtmlEl[] | string | (HtmlEl | string)[] | undefined
 ): Text | DocumentFragment | HTMLElement {
     if (block === undefined) {
         return document.createTextNode('');
